+++ src/task/task.ts
@@ -32,7 +32,7 @@ export class TaskError extends Error {}
 // Anything can be assigned to a task's `args`. It's just a mutable POJO dictionary.
 // Downstream tasks will see args that upstream events mutate,
 // unless `invokeTask` is called with modified args.
-// Upstream tasks can use hooks to respond to downstream events and values.
+// Upstream tasks can use listeners to respond to downstream events and values.
 // It's a beautiful mutable spaghetti mess. cant get enough
 // The raw CLI ares are handled by `mri` - https://github.com/lukeed/mri
 export interface Args {